+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/RotHeart.java
@@ -21,6 +21,7 @@
 
 package com.shatteredpixel.shatteredpixeldungeon.actors.mobs;
 
+import com.shatteredpixel.shatteredpixeldungeon.Challenges;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Blob;
@@ -93,7 +94,8 @@ public class RotHeart extends Mob {
 	@Override
 	public void die(Object cause) {
 		super.die(cause);
-		Dungeon.level.drop( new Rotberry.Seed(), pos()).sprite.drop();
+		if	(!isAlive())
+			Dungeon.level.drop( new Rotberry.Seed(), pos()).sprite.drop();
 	}
 
 	@Override
